+++ packages/swirl-components/src/components/flip-popover/flip-popover.tsx
@@ -3,6 +3,7 @@ import {
   computePosition,
   ComputePositionReturn,
   flip,
+  Placement,
   shift,
 } from "@floating-ui/dom";
 import {
@@ -11,12 +12,13 @@ import {
   h,
   Host,
   Listen,
+  Method,
   Prop,
   State,
 } from "@stencil/core";
-import classnames from "classnames";
-import { querySelectorAllDeep } from "../../utils";
 import { disableBodyScroll, enableBodyScroll } from "body-scroll-lock";
+import classnames from "classnames";
+import { isMobileViewport, querySelectorAllDeep } from "../../utils";
 
 /**
  * @slot slot - The popover content.
@@ -30,6 +32,7 @@ export class FlipPopover {
   @Element() el: HTMLElement;
 
   @Prop() label!: string;
+  @Prop() placement?: Placement = "bottom-start";
   @Prop() popoverId!: string;
   @Prop() trigger!: string;
 
@@ -53,25 +56,36 @@ export class FlipPopover {
     enableBodyScroll(this.scrollContainer);
   }
 
-  @Listen("focusout", { target: "window" })
-  onWindowFocusout(event: FocusEvent) {
-    if (!this.active) {
-      return;
-    }
+  @Listen("click", { target: "window" })
+  onWindowClick(event: MouseEvent) {
+    const target = event.target as HTMLElement;
 
-    const target = event.relatedTarget as HTMLElement;
-
-    const popoverLostFocus =
-      target === null ||
-      (!this.el.contains(target) &&
-        !this.childMenuItems.some((item) => target.shadowRoot?.contains(item)));
-
-    if (popoverLostFocus) {
+    if (!this.el.contains(target)) {
       this.close();
     }
   }
 
-  close = () => {
+  onFocusOut = (event: FocusEvent) => {
+    if (!this.active) {
+      return;
+    }
+
+    const target =
+      (event.relatedTarget as HTMLElement) || (event.target as HTMLElement);
+
+    const popoverLostFocus = !this.el.contains(target);
+
+    if (popoverLostFocus) {
+      this.close();
+    }
+  };
+
+  /**
+   * Close the popover.
+   * @returns
+   */
+  @Method()
+  public async close() {
     if (this.closing) {
       return;
     }
@@ -90,10 +104,15 @@ export class FlipPopover {
 
     this.unlockBodyScroll();
 
-    this.triggerEl?.focus();
-  };
+    this.getNativeTriggerElement()?.focus();
+  }
 
-  open = () => {
+  /**
+   * Open the popover.
+   * @returns
+   */
+  @Method()
+  public async open() {
     this.active = true;
 
     this.updateChildMenuItems();
@@ -122,9 +141,11 @@ export class FlipPopover {
 
       this.lockBodyScroll();
     });
-  };
+  }
 
-  toggle = () => {
+  toggle = (event: Event) => {
+    event.stopPropagation();
+
     if (this.active) {
       this.close();
     } else {
@@ -133,14 +154,7 @@ export class FlipPopover {
   };
 
   private connectTrigger() {
-    const triggerComponent = querySelectorAllDeep(
-      document.body,
-      `#${this.trigger}`
-    )[0];
-
-    this.triggerEl = (triggerComponent?.children[0] ||
-      triggerComponent?.shadowRoot?.children[0] ||
-      triggerComponent) as HTMLElement;
+    this.triggerEl = querySelectorAllDeep(document.body, `#${this.trigger}`)[0];
 
     if (!Boolean(this.triggerEl)) {
       return;
@@ -149,18 +163,18 @@ export class FlipPopover {
     this.triggerEl.addEventListener("click", this.toggle);
   }
 
+  private getNativeTriggerElement() {
+    return this.triggerEl.tagName.startsWith("FLIP-")
+      ? ((this.triggerEl?.children[0] ||
+          this.triggerEl?.shadowRoot?.children[0] ||
+          this.triggerEl) as HTMLElement)
+      : this.triggerEl;
+  }
+
   private onKeydown = (event: KeyboardEvent) => {
     if (event.code === "Escape" && this.active) {
       this.close();
     }
-  };
-
-  private onContentClick = () => {
-    if (!this.active) {
-      return;
-    }
-
-    this.close();
   };
 
   private updateTriggerAttributes() {
@@ -168,9 +182,11 @@ export class FlipPopover {
       return;
     }
 
-    this.triggerEl.setAttribute("aria-controls", this.popoverId);
-    this.triggerEl.setAttribute("aria-expanded", String(this.active));
-    this.triggerEl.setAttribute("aria-haspopup", "dialog");
+    const nativeTriggerEl = this.getNativeTriggerElement();
+
+    nativeTriggerEl.setAttribute("aria-controls", this.popoverId);
+    nativeTriggerEl.setAttribute("aria-expanded", String(this.active));
+    nativeTriggerEl.setAttribute("aria-haspopup", "dialog");
   }
 
   private updateChildMenuItems() {
@@ -178,7 +194,7 @@ export class FlipPopover {
   }
 
   private reposition = async () => {
-    const mobile = !window.matchMedia("(min-width: 768px)").matches;
+    const mobile = isMobileViewport();
 
     if (!Boolean(this.triggerEl) || !Boolean(this.contentContainer)) {
       return;
@@ -194,14 +210,14 @@ export class FlipPopover {
       this.contentContainer,
       {
         middleware: [shift(), flip()],
-        placement: "bottom-start",
+        placement: this.placement,
         strategy: "fixed",
       }
     );
   };
 
   private lockBodyScroll() {
-    const mobile = !window.matchMedia("(min-width: 768px)").matches;
+    const mobile = isMobileViewport();
 
     if (!mobile) {
       return;
@@ -214,6 +230,10 @@ export class FlipPopover {
     enableBodyScroll(this.scrollContainer);
   }
 
+  private onCloseButtonClick = () => {
+    this.close();
+  };
+
   render() {
     const className = classnames("popover", {
       "popover--closing": this.closing,
@@ -222,13 +242,12 @@ export class FlipPopover {
     });
 
     return (
-      <Host id={this.popoverId}>
+      <Host id={this.popoverId} onFocusout={this.onFocusOut}>
         <div class={className} onKeyDown={this.onKeydown}>
           <div
             aria-hidden={!this.active ? "true" : "false"}
             aria-label={this.label}
             class="popover__content"
-            onClick={this.onContentClick}
             role="dialog"
             tabindex="-1"
             ref={(el) => (this.contentContainer = el)}
@@ -246,7 +265,10 @@ export class FlipPopover {
             </div>
           </div>
           {this.active && (
-            <div class="popover__backdrop" onClick={this.close}></div>
+            <div
+              class="popover__backdrop"
+              onClick={this.onCloseButtonClick}
+            ></div>
           )}
         </div>
       </Host>